+++ tests/py/apps/searchengine/meta.py
@@ -3,7 +3,7 @@ from random import randint
 from stdnet import odm, QuerySetError
 from stdnet.utils import test
 from stdnet.odm.search import UpdateSE
-from stdnet.utils import test, to_string, range, populate
+from stdnet.utils import test, to_string, range
 from stdnet.apps.searchengine import SearchEngine, processors
 
 from examples.wordsearch.basicwords import basic_english_words
@@ -50,12 +50,43 @@ NAMES = {'maurice':('MRS', None),
          'Nowhere':('NR', None),
          'Tux':('TKS', None)}
 
-
-NUM_WORDS = 40
-WORDS_GROUPS = lambda size : (' '.join(populate('choice', NUM_WORDS,\
-                              choice_from = basic_english_words))\
-                               for i in range(size))
     
+class SeearchData(test.DataGenerator):
+    sizes = {'tiny': (10, 10),
+             'small': (50, 20),
+             'normal': (500, 30),
+             'big': (10000, 40),
+             'huge': (1000000, 50)}
+    
+    def generate(self):
+        size, num = self.size
+        self.names = self.populate('choice', size=size,
+                                   choice_from=basic_english_words)
+        self.groups = []
+        self.empty = size*['']
+        for i in range(size):
+            text = ' '.join(self.populate('choice', num,\
+                                          choice_from=basic_english_words))
+            self.groups.append(text)
+    
+    def make_items(self, test, content=False, related=None):
+        '''Bulk creation of Item for testing search engine. Return a set
+of words which have been included in the Items.'''
+        session = test.session()
+        words = set()
+        contents = self.groups if content else self.empty
+        with session.begin() as t:
+            for name, content in zip(self.names, contents):
+                if len(name) > 3:
+                    words.add(name)
+                    if content:
+                        words.update(content.split())
+                    t.add(Item(name=name, counter=randint(0,10),
+                               content=content, related=related))
+        yield t.on_result
+        test.words = words
+        yield test.words
+        
 
 class SearchMixin(object):
     '''Mixin for testing the search engine. No tests implemented here,
@@ -65,6 +96,7 @@ below will derive from this class.'''
     metaphone = True
     stemming = True
     models = (Item, RelatedItem)
+    data_cls = SeearchData
     
     @classmethod
     def after_setup(cls):
@@ -78,34 +110,8 @@ below will derive from this class.'''
     
     def make_item(self, name='python', counter=10, content=None, related=None):
         content = content if content is not None else python_content
-        with self.session().begin() as t:
-            item = t.add(Item(name=name, counter=counter,
-                              content=content, related=related))
-        yield t.on_result
-        yield item
-    
-    @classmethod
-    def make_items(cls, num=30, content=False, related=None):
-        '''Bulk creation of Item for testing search engine. Return a set
-of words which have been included in the Items.'''
-        names = populate('choice', num, choice_from=basic_english_words)
-        session = cls.session()
-        words = set()
-        if content:
-            contents = WORDS_GROUPS(num)
-        else:
-            contents = ['']*num
-        with session.begin() as t:
-            for name, content in zip(names, contents):
-                if len(name) > 3:
-                    words.add(name)
-                    if content:
-                        words.update(content.split())
-                    t.add(Item(name=name, counter=randint(0,10),
-                               content=content, related=related))
-        yield t.on_result
-        cls.words = words
-        yield cls.words
+        return self.mapper.item.new(name=name, counter=counter, content=content,
+                                    related=related)
     
     def simpleadd(self, name='python', counter=10, content=None, related=None):
         models = self.mapper
@@ -114,8 +120,8 @@ of words which have been included in the Items.'''
         wis = yield engine.worditems(item).all()
         self.assertTrue(wis)
         session = self.session()
-        objets = yield self.multi_async((wi.object(session) for wi in wis))
-        for object in objets:
+        objects = yield self.multi_async((wi.object(session) for wi in wis))
+        for object in objects:
             self.assertEqual(object, item)
         yield item, wis
         
@@ -182,7 +188,7 @@ class TestCoverageBaseClass(test.TestCase):
         self.assertRaises(NotImplementedError, e.flush)
         self.assertRaises(NotImplementedError, e.add_item, None, None, None)
         self.assertRaises(NotImplementedError, e.remove_item, None, None, None)
-        self.assertRaises(NotImplementedError, e.session)
+        self.assertRaises(AttributeError, e.session)
         self.assertEqual(e.split_text('ciao luca'), ['ciao','luca'])
     
     def testItemFieldIterator(self):