+++ CHANGELOG.rst
@@ -50,7 +50,7 @@ Ver. 0.8 - Development
   Check the :ref:`asynchronous tutorial <tutorial-asynchronous>` for information.
 * Refactored test suite to handle asynchronous and synchronous connections.
 * Redis ``odm`` script is able to fix unique-keys/ids conflicts.
-* **538 regression tests** with **93%** coverage.
+* **542 regression tests** with **93%** coverage.
   
 .. _vers07: